requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9
defusedxml==0.6.0
https://github.com/QizaiMing/django-custom/archive/stable/2.2.x.zip
django-model-utils==4.0.0
django-notifications-hq==1.6.0
ecdsa==0.15
filestack-python==3.1.0
gunicorn==20.0.4
idna==2.8
jsonfield==3.0.0
oauthlib==3.1.0
Pillow==7.1.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.12.0
python-jose==3.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
swapper==1.1.2.post1
trafaret==1.2.0
urllib3==1.25.8
whitenoise==5.2.0
